requirements.txt
discord.py
selenium
pillow